SDESystem

System Constructors

ModelingToolkit.SDESystem — Type

struct SDESystem <: ModelingToolkit.AbstractODESystem

A system of stochastic differential equations.

Fields

eqs

The expressions defining the drift term.

noiseeqs

The expressions defining the diffusion term.

iv

Independent variable.

states

Dependent (state) variables. Must not contain the independent variable.

ps

Parameter variables. Must not contain the independent variable.

var_to_name

Array variables.

ctrls

Control parameters (some subset of ps).

observed

Observed states.

tgrad

Time-derivative matrix. Note: this field will not be defined until calculate_tgrad is called on the system.

jac

Jacobian matrix. Note: this field will not be defined until calculate_jacobian is called on the system.

ctrl_jac

Control Jacobian matrix. Note: this field will not be defined until calculate_control_jacobian is called on the system.

Wfact

Wfact matrix. Note: this field will not be defined until generate_factorized_W is called on the system.

Wfact_t

Wfact_t matrix. Note: this field will not be defined until generate_factorized_W is called on the system.

name

Name: the name of the system

systems

Systems: the internal systems. These are required to have unique names.

defaults

defaults: The default values to use when initial conditions and/or parameters are not supplied in ODEProblem.

connector_type

type: type of the system

Example

using ModelingToolkit

@parameters σ ρ β
@variables t x(t) y(t) z(t)
D = Differential(t)

eqs = [D(x) ~ σ*(y-x),
       D(y) ~ x*(ρ-z)-y,
       D(z) ~ x*y - β*z]

noiseeqs = [0.1*x,
            0.1*y,
            0.1*z]

@named de = SDESystem(eqs,noiseeqs,t,[x,y,z],[σ,ρ,β])

To convert an ODESystem to an SDESystem directly:

ode = ODESystem(eqs,t,[x,y,z],[σ,ρ,β])
sde = SDESystem(ode, noiseeqs)

Composition and Accessor Functions

get_eqs(sys) or equations(sys): The equations that define the SDE.

get_states(sys) or states(sys): The set of states in the SDE.

get_ps(sys) or parameters(sys): The parameters of the SDE.

get_iv(sys): The independent variable of the SDE.

Transformations

Missing docstring.

Missing docstring for structural_simplify. Check Documenter's build log for details.

Missing docstring.

Missing docstring for alias_elimination. Check Documenter's build log for details.

Analyses

Applicable Calculation and Generation Functions

calculate_jacobian
calculate_tgrad
calculate_factorized_W
generate_jacobian
generate_tgrad
generate_factorized_W
jacobian_sparsity

Problem Constructors

SciMLBase.SDEFunction — Type

SDEFunction{iip,F,G,TMM,Ta,Tt,TJ,JVP,VJP,JP,SP,TW,TWt,TPJ,GG,S,O,TCV} <: AbstractSDEFunction{iip}

A representation of an SDE function f, defined by:

\[M du = f(u,p,t)dt + g(u,p,t) dW \]

and all of its related functions, such as the Jacobian of f, its gradient with respect to time, and more. For all cases, u0 is the initial condition, p are the parameters, and t is the independent variable.

Constructor

SDEFunction{iip,recompile}(f,g;
                           mass_matrix=I,
                           analytic=nothing,
                           tgrad=nothing,
                           jac=nothing,
                           jvp=nothing,
                           vjp=nothing,
                           ggprime = nothing,
                           jac_prototype=nothing,
                           sparsity=jac_prototype,
                           paramjac = nothing,
                           syms = nothing,
                           indepsym = nothing,
                           colorvec = nothing)

Note that only the function f itself is required. This function should be given as f!(du,u,p,t) or du = f(u,p,t). See the section on iip for more details on in-place vs out-of-place handling.

All of the remaining functions are optional for improving or accelerating the usage of f. These include:

mass_matrix: the mass matrix M represented in the ODE function. Can be used to determine that the equation is actually a differential-algebraic equation (DAE) if M is singular. Note that in this case special solvers are required, see the DAE solver page for more details: https://diffeq.sciml.ai/stable/solvers/dae_solve/. Must be an AbstractArray or an AbstractSciMLOperator.

analytic(u0,p,t): used to pass an analytical solution function for the analytical solution of the ODE. Generally only used for testing and development of the solvers.

tgrad(dT,u,p,t) or dT=tgrad(u,p,t): returns $\frac{\partial f(u,p,t)}{\partial t}$

jac(J,u,p,t) or J=jac(u,p,t): returns $\frac{df}{du}$

jvp(Jv,v,u,p,t) or Jv=jvp(v,u,p,t): returns the directional derivative$\frac{df}{du} v$

vjp(Jv,v,u,p,t) or Jv=vjp(v,u,p,t): returns the adjoint derivative$\frac{df}{du}^\ast v$

ggprime(J,u,p,t) or J = ggprime(u,p,t): returns the Milstein derivative $\frac{dg(u,p,t)}{du} g(u,p,t)$

jac_prototype: a prototype matrix matching the type that matches the Jacobian. For example, if the Jacobian is tridiagonal, then an appropriately sized Tridiagonal matrix can be used as the prototype and integrators will specialize on this structure where possible. Non-structured sparsity patterns should use a SparseMatrixCSC with a correct sparsity pattern for the Jacobian. The default is nothing, which means a dense Jacobian.

paramjac(pJ,u,p,t): returns the parameter Jacobian $\frac{df}{dp}$.

syms: the symbol names for the elements of the equation. This should match u0 in size. For example, if u0 = [0.0,1.0] and syms = [:x, :y], this will apply a canonical naming to the values, allowing sol[:x] in the solution and automatically naming values in plots.

indepsym: the canonical naming for the independent variable. Defaults to nothing, which internally uses t as the representation in any plots.

colorvec: a color vector according to the SparseDiffTools.jl definition for the sparsity pattern of the jac_prototype. This specializes the Jacobian construction when using finite differences and automatic differentiation to be computed in an accelerated manner based on the sparsity pattern. Defaults to nothing, which means a color vector will be internally computed on demand when required. The cost of this operation is highly dependent on the sparsity pattern.

iip: In-Place vs Out-Of-Place

For more details on this argument, see the ODEFunction documentation.

recompile: Controlling Compilation and Specialization

For more details on this argument, see the ODEFunction documentation.

Fields

The fields of the ODEFunction type directly match the names of the inputs.

SciMLBase.SDEProblem — Type

Defines an stochastic differential equation (SDE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/sde_types/

Mathematical Specification of a SDE Problem

To define an SDE Problem, you simply need to give the forcing function f, the noise function g, and the initial condition u₀ which define an SDE:

\[du = f(u,p,t)dt + Σgᵢ(u,p,t)dWⁱ\]

f and g should be specified as f(u,p,t) and g(u,p,t) respectively, and u₀ should be an AbstractArray whose geometry matches the desired geometry of u. Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well. A vector of gs can also be defined to determine an SDE of higher Ito dimension.

Problem Type

Wraps the data which defines an SDE problem

\[u = f(u,p,t)dt + Σgᵢ(u,p,t)dWⁱ\]

with initial condition u0.

Constructors

SDEProblem(f::SDEFunction,g,u0,tspan,p=NullParameters();noise=WHITE_NOISE,noise_rate_prototype=nothing)

SDEProblem{isinplace}(f,g,u0,tspan,p=NullParameters();noise=WHITE_NOISE,noise_rate_prototype=nothing) : Defines the SDE with the specified functions. The default noise is WHITE_NOISE. isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the DiffEqFunctions page.

Fields

f: The drift function in the SDE.

g: The noise function in the SDE.

u0: The initial condition.

tspan: The timespan for the problem.

p: The optional parameters for the problem. Defaults to NullParameters.

noise: The noise process applied to the noise upon generation. Defaults to Gaussian white noise. For information on defining different noise processes, see the noise process documentation page

noise_rate_prototype: A prototype type instance for the noise rates, that is the output g. It can be any type which overloads A_mul_B! with itself being the middle argument. Commonly, this is a matrix or sparse matrix. If this is not given, it defaults to nothing, which means the problem should be interpreted as having diagonal noise.

kwargs: The keyword arguments passed onto the solves.

Example Problems

Examples problems can be found in DiffEqProblemLibrary.jl.

To use a sample problem, such as prob_sde_linear, you can do something like:

#] add DiffEqProblemLibrary
using DiffEqProblemLibrary.SDEProblemLibrary
# load problems
SDEProblemLibrary.importsdeproblems()
prob = SDEProblemLibrary.prob_sde_linear
sol = solve(prob)