ModelingToolkitStandardLibrary: Blocks

Utility Blocks

ModelingToolkitStandardLibrary.Blocks.RealInput — Function

RealInput(;name, nin=1, u_start=nin > 1 ? 0.0 : zeros(nin))

Connector with one input signal of type Real.

Parameters:

nin: Number of inputs

u_start: Initial value for u

States:

u: Value of of the connector; if nin=1 this is a scalar

ModelingToolkitStandardLibrary.Blocks.RealOutput — Function

RealOutput(;name, nout=1, u_start=nout > 1 ? 0.0 : zeros(nout))

Connector with one output signal of type Real.

Parameters:

nout: Number of inputs

u_start: Initial value for u

States:

u: Value of of the connector; if nout=1 this is a scalar

ModelingToolkitStandardLibrary.Blocks.SISO — Function

SISO(;name, u_start=0.0, y_start=0.0)

Single Input Single Output continuous control block.

Parameters:

u_start: Initial value for the input

y_start: Initial value for the output

Math Blocks

ModelingToolkitStandardLibrary.Blocks.Gain — Function

Gain(k; name)

Output the product of a gain value with the input signal.

Parameters:

k: Scalar gain

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.MatrixGain — Function

MatrixGain(K::AbstractArray; name)

Output the product of a gain matrix with the input signal vector.

Parameters:

K: Matrix gain

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.Sum — Function

Sum(n::Int; name)

Output the sum of the elements of the input port vector.

Parameters:

n: Input port dimension

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.Feedback — Function

Feedback(;name)

Output difference between reference input (input1) and feedback input (input2).

Connectors:

input1

input2

output

ModelingToolkitStandardLibrary.Blocks.Add — Function

Add(;name, k1=1, k2=1)

Output the sum of the two scalar inputs.

Parameters:

k1: Gain for first input

k2: Gain for second input

Connectors:

input1

input2

output

ModelingToolkitStandardLibrary.Blocks.Add3 — Function

Add(;name, k1=1, k2=1,k3=1)

Output the sum of the three scalar inputs.

Parameters:

k1: Gain for first input

k2: Gain for second input

k3: Gain for third input

Connectors:

input1

input2

input3

output

ModelingToolkitStandardLibrary.Blocks.Product — Function

Product(;name)

Output product of the two inputs.

Connectors:

input1

input2

output

ModelingToolkitStandardLibrary.Blocks.Division — Function

Division(;name)

Output first input divided by second input.

Connectors:

input1

input2

output

ModelingToolkitStandardLibrary.Blocks.StaticNonLinearity — Function

StaticNonLinearity(func ;name)

Applies the given function to the input.

If the given function is not composed of simple core methods (e.g. sin, abs, ...), it has to be registered via @register_symbolic func(u)

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.Abs — Function

Abs(;name)

Output the absolute value of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Sign — Function

Sign(;name)

Output the sign of the input

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Sqrt — Function

Sqrt(;name)

Output the square root of the input (input >= 0 required).

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Sin — Function

Sin(;name)

Output the sine of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Cos — Function

Cos(;name)

Output the cosine of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Tan — Function

Tan(;name)

Output the tangent of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Asin — Function

Asin(;name)

Output the arc sine of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Acos — Function

Acos(;name)

Output the arc cosine of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Atan — Function

Atan(;name)

Output the arc tangent of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Atan2 — Function

Atan2(;name)

Output the arc tangent of the input.

Connectors:

input1

input2

output

ModelingToolkitStandardLibrary.Blocks.Sinh — Function

Sinh(;name)

Output the hyperbolic sine of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Cosh — Function

Cosh(;name)

Output the hyperbolic cosine of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Tanh — Function

Tanh(;name)

Output the hyperbolic tangent of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Exp — Function

Exp(;name)

Output the exponential (base e) of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Log — Function

Log(;name)

Output the natural (base e) logarithm of the input.

Connectors:

See StaticNonLinearity

ModelingToolkitStandardLibrary.Blocks.Log10 — Function

Log10(;name)

Output the base 10 logarithm of the input.

Connectors:

See StaticNonLinearity

Source Blocks

ModelingToolkitStandardLibrary.Blocks.Constant — Function

Generate constant signal.

Parameters:

k: Constant output value

Connectors:

output

ModelingToolkitStandardLibrary.Blocks.Sine — Function

Generate sine signal.

Parameters:

frequency: [Hz] Frequency of sine wave

amplitude: Amplitude of sine wave

phase: [rad] Phase of sine wave

offset: Offset of output signal

start_time: [s] Output y = offset for t < start_time

smooth: If true, returns a smooth wave. Defaults to false It uses a smoothing factor of δ=1e-5

Connectors:

output

Missing docstring.

Missing docstring for Cosine. Check Documenter's build log for details.

ModelingToolkitStandardLibrary.Blocks.ContinuousClock — Function

Generate current time signal.

Parameters:

offset: Offset of output signal

start_time: [s] Output y = offset for t < start_time

Connectors:

output

ModelingToolkitStandardLibrary.Blocks.Ramp — Function

Generate ramp signal.

Parameters:

height: Height of ramp

duration: [s] Duration of ramp (= 0.0 gives a Step)

offset: Offset of output signal

start_time: [s] Output y = offset for t < start_time

smooth: If true, returns a smooth wave. Defaults to false It uses a smoothing factor of δ=1e-5

Connectors:

output

ModelingToolkitStandardLibrary.Blocks.Step — Function

Generate step signal.

Parameters:

height: Height of step

offset: Offset of output signal

start_time: [s] Output y = offset for t < start_time

smooth: If true, returns a smooth wave. Defaults to false It uses a smoothing factor of δ=1e-5

Connectors:

output

ModelingToolkitStandardLibrary.Blocks.ExpSine — Function

Generate exponentially damped sine signal.

Parameters:

frequency: [Hz] Frequency of sine wave

amplitude: Amplitude of sine wave

damping: [1/s] Damping coefficient of sine wave

phase: [rad] Phase of sine wave

offset: Offset of output signal

start_time: [s] Output y = offset for t < start_time

smooth: If true, returns a smooth wave. Defaults to false It uses a smoothing factor of δ=1e-5

Connectors:

output

Nonlinear Blocks

ModelingToolkitStandardLibrary.Blocks.Limiter — Function

Limiter(;name, y_max, y_min=y_max > 0 ? -y_max : -Inf)

Limit the range of a signal.

Parameters:

y_max: Maximum of output signal

y_min: Minimum of output signal

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.DeadZone — Function

DeadZone(; name, u_max, u_min=-u_max)

The DeadZone block defines a region of zero output. If the input is within u_min ... u_max, the output is zero. Outside of this zone, the output is a linear function of the input with a slope of 1.

       y▲
        │     /
        │    /
  u_min │   /
─────|──┼──|───────► u
    /   │   u_max
   /    │
  /     │

Parameters:

u_max: Upper limit of dead zone

u_min: Lower limit of dead zone

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.SlewRateLimiter — Function

SlewRateLimiter(;name, rising=1, falling=-rising, Td=0.001, y_start=0.0)

Limits the slew rate of a signal.

Parameters:

rising: Maximum rising slew rate

falling: Maximum falling slew rate

Td: [s] Derivative time constant

Connectors:

input

output

Continuous Blocks

ModelingToolkitStandardLibrary.Blocks.Integrator — Function

Integrator(;name, k=1, x_start=0.0)

Outputs y = ∫k*u dt, corresponding to the transfer function 1/s.

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.Derivative — Function

Derivative(; name, k=1, T, x_start=0.0)

Outputs an approximate derivative of the input. The transfer function of this block is

k       k     
─ - ──────────
T    2 ⎛    1⎞
    T ⋅⎜s + ─⎟
       ⎝    T⎠

and a state-space realization is given by ss(-1/T, 1/T, -k/T, k/T) where T is the time constant of the filter. A smaller T leads to a more ideal approximation of the derivative.

Parameters:

k: Gain

T: [s] Time constants (T>0 required; T=0 is ideal derivative block)

x_start: Initial value of state

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.FirstOrder — Function

FirstOrder(; name, k=1, T, x_start=0.0)

A first-order filter with a single real pole in s = -T and gain k. The transfer function is given by Y(s)/U(s) =

   k   
───────
sT + 1

Parameters:

k: Gain

T: [s] Time constants (T>0 required)

x_start: Initial value of state

Connectors:

input

output

See also SecondOrder

ModelingToolkitStandardLibrary.Blocks.SecondOrder — Function

SecondOrder(; name, k=1, w, d, x_start=0.0, xd_start=0.0)

A second-order filter with gain k, a bandwidth of w rad/s and relative damping d. The transfer function is given by Y(s)/U(s) =

      k*w^2   
─────────────────
s² + 2d*w*s + w^2

Critical damping corresponds to d=1, which yields the fastest step response without overshoot, d < 1results in an under-damped filter whiled > 1results in an over-damped filter.d = 1/√2` corresponds to a Butterworth filter of order 2 (maximally flat frequency response).

Parameters:

k: Gain

w: Angular frequency

d: Damping

x_start: Initial value of state (output)

xd_start: Initial value of derivative of state (output)

Connectors:

input

output

ModelingToolkitStandardLibrary.Blocks.StateSpace — Function

StateSpace(A, B, C, D=0; x_start=zeros(size(A,1)), name)

A linear, time-invariant state-space system on the form.

ẋ = Ax + Bu
y = Cx + Du

Transfer functions can also be simulated by converting them to a StateSpace form.

ModelingToolkitStandardLibrary.Blocks.PI — Function

PI(;name, k=1, T, x_start=0.0)

Textbook version of a PI-controller without actuator saturation and anti-windup measure.

Parameters:

k: Gain

T: [s] Integrator time constant (T>0 required)

x_start: Initial value for the integrator

Connectors:

err_input

ctr_output

See also LimPI

ModelingToolkitStandardLibrary.Blocks.LimPI — Function

LimPI(;name, k=1, T, u_max=1, u_min=-u_max, Ta)

Text-book version of a PI-controller with actuator saturation and anti-windup measure.

Parameters:

k: Gain

T: [s] Integrator time constant (T>0 required)

Ta: [s] Tracking time constant (Ta>0 required)

x_start: Initial value for the integrator

Connectors:

err_input

ctr_output

ModelingToolkitStandardLibrary.Blocks.PID — Function

PID(;name, k=1, Ti=false, Td=false, Nd=10, xi_start=0, xd_start=0)

Text-book version of a PID-controller without actuator saturation and anti-windup measure.

Parameters:

k: Gain

Ti: [s] Integrator time constant (Ti>0 required). If set to false no integral action is used.

Td: [s] Derivative time constant (Td>0 required). If set to false no derivative action is used.

Nd: [s] Time constant for the derivative approximation (Nd>0 required; Nd=0 is ideal derivative).

x_start: Initial value for the integrator.

xd_start: Initial value for the derivative state.

Connectors:

err_input

ctr_output

See also LimPID

ModelingToolkitStandardLibrary.Blocks.LimPID — Function

LimPID(; k, Ti=false, Td=false, wp=1, wd=1, Ni, Nd=12, u_max=Inf, u_min=-u_max, gains = false, name)

Proportional-Integral-Derivative (PID) controller with output saturation, set-point weighting and integrator anti-windup.

The equation for the control signal is roughly

k(ep + 1/Ti * ∫e + 1/Td * d/dt(ed))
e = u_r - u_y
ep = wp*u_r - u_y
ed = wd*u_r - u_y

where the transfer function for the derivative includes additional filtering, see ? Derivative for more details.

Parameters:

k: Proportional gain

Ti: [s] Integrator time constant. Set to false to turn off integral action.

Td: [s] Derivative time constant. Set to false to turn off derivative action.

wp: [0,1] Set-point weighting in the proportional part.

wd: [0,1] Set-point weighting in the derivative part.

Nd: [1/s] Derivative limit, limits the derivative gain to Nd/Td. Reasonable values are ∈ [8, 20]. A higher value gives a better approximation of an ideal derivative at the expense of higher noise amplification.

Ni: Ni*Ti controls the time constant Ta of anti-windup tracking. A common (default) choice is Ta = √(Ti*Td) which is realized by Ni = √(Td / Ti). Anti-windup can be effectively turned off by setting Ni = Inf.

gains: Ifgains = true,TiandTdwill be interpreted as gains with a fundamental PID transfer function on parallel formki=Ti, kd=Td, k + ki/s + kd*s`

Connectors:

reference

measurement

ctr_output