ODE Tableaus

Explicit Runge-Kutta Methods

constructEuler - Euler's 1st order method.

constructHuen() Huen's order 2 method.

constructRalston() - Ralston's order 2 method.

constructSSPRK22() - Explicit SSP method of order 2 using 2 stages.

constructKutta3 - Kutta's classic 3rd order method.

constructSSPRK33() - Explicit SSP method of order 3 using 3 stages.

constructSSPRK43() - Explicit SSP method of order 3 using 4 stages.

constructRK4 - The classic 4th order "Runge-Kutta" method.

constructRK438Rule - The classic 4th order "3/8th's Rule" method.

constructSSPRK104() - Explicit SSP method of order 4 using 10 stages.

constructBogakiShampine3() - Bogakai-Shampine's 2/3 method.

constructRKF4() - Runge-Kutta-Fehlberg 3/4.

constructRKF5() - Runge-Kutta-Fehlberg 4/5.

constructRungeFirst5() - Runge's first 5th order method.

constructCassity5() - Cassity's 5th order method.

constructLawson5() - Lawson's 5th order method.

constructLutherKonen5 - Luther-Konen's first 5th order method.

constructLutherKonen52() - Luther-Konen's second 5th order method.

constructLutherKonen53() - Luther-Konen's third 5th order method.

constructPapakostasPapaGeorgiou5() - Papakostas and PapaGeorgiou more stable order 5 method.

constructPapakostasPapaGeorgiou52() - Papakostas and PapaGeorgiou more efficient order 5 method.

constructTsitouras5() - Tsitouras's order 5 method.

constructBogakiShampine5() - Bogaki and Shampine's Order 5 method.

constructSharpSmart5() - Sharp and Smart's Order 5 method.

constructCashKarp() - Cash-Karp method 4/5.

constructDormandPrince() - Dormand-Prince 4/5.

constructButcher6() - Butcher's first order 6 method.

constructButcher62() - Butcher's second order 6 method.

constructButcher63() - Butcher's third order 6 method.

constructDormandPrince6() - Dormand-Prince's 5/6 method.

constructSharpVerner6() Sharp-Verner's 5/6 method.

constructVerner916() - Verner's more efficient order 6 method (1991).

constructVerner9162() - Verner's second more efficient order 6 method (1991).

constructVernerRobust6() - Verner's "most robust" order 6 method.

constructVernerEfficient6() - Verner's "most efficient" order 6 method.

constructPapakostas6() - Papakostas's order 6 method.

constructLawson6() - Lawson's order 6 method.

constructTsitourasPapakostas6() - Tsitouras and Papakostas's order 6 method.

constructDormandLockyerMcCorriganPrince6() - the Dormand-Lockyer-McCorrigan-Prince order 6 method.

constructTanakaKasugaYamashitaYazaki6A() - Tanaka-Kasuga-Yamashita-Yazaki order 6 method A.

constructTanakaKasugaYamashitaYazaki6B() - Tanaka-Kasuga-Yamashita-Yazaki order 6 method B.

constructTanakaKasugaYamashitaYazaki6C() - Tanaka-Kasuga-Yamashita-Yazaki order 6 method C.

constructTanakaKasugaYamashitaYazaki6D() - Tanaka-Kasuga-Yamashita-Yazaki order 6 method D.

constructMikkawyEisa() - Mikkawy and Eisa's order 6 method.

constructChummund6() - Chummund's first order 6 method.

constructChummund62() - Chummund's second order 6 method.

constructHuta6() - Huta's first order 6 method.

constructHuta62() - Huta's second order 6 method.

constructVerner6() - An old order 6 method attributed to Verner.

constructDverk() - The classic DVERK algorithm attributed to Verner.

constructClassicVerner6() - A classic Verner order 6 algorithm (1978).

constructButcher7() - Butcher's order 7 algorithm.

constructClassicVerner7()- A classic Verner order 7 algorithm (1978).

constructVernerRobust7() - Verner's "most robust" order 7 algorithm.

constructTanakaYamashitaStable7() - Tanaka-Yamashita more stable order 7 algorithm.

constructTanakaYamashitaEfficient7() - Tanaka-Yamashita more efficient order 7 algorithm.

constructSharpSmart7() - Sharp-Smart's order 7 algorithm.

constructSharpVerner7() - Sharp-Verner's order 7 algorithm.

constructVerner7() - Verner's "most efficient" order 7 algorithm.

constructVernerEfficient7() - Verner's "most efficient" order 7 algorithm.

constructClassicVerner8() - A classic Verner order 8 algorithm (1978).

constructCooperVerner8() - Cooper-Verner's first order 8 algorithm.

constructCooperVerner82() - Cooper-Verner's second order 8 algorithm.

constructTsitourasPapakostas8() - Tsitouras-Papakostas order 8 algorithm.

constructdverk78() - The classic order 8 DVERK algorithm.

constructEnrightVerner8() - Enright-Verner order 8 algorithm.

constructCurtis8() - Curtis' order 8 algorithm.

constructVerner8() - Verner's "most efficient" order 8 algorithm.

constructRKF8() - Runge-Kutta-Fehlberg Order 7/8 method.

constructDormandPrice8() - Dormand-Prince Order 7/8 method.

constructDormandPrince8_64bit() - Dormand-Prince Order 7/8 method. Coefficients are rational approximations good for 64 bits.

constructVernerRobust9() - Verner's "most robust" order 9 method.

constructVernerEfficient9() - Verner's "most efficient" order 9 method.

constructSharp9() - Sharp's order 9 method.

constructTsitouras9() - Tsitouras's first order 9 method.

constructTsitouras92() - Tsitouras's second order 9 method.

constructCurtis10() - Curtis' order 10 method.

constructOno10() - Ono's order 10 method.

constructFeagin10Tableau() - Feagin's order 10 method.

constructCurtis10() - Curtis' order 10 method.

constructBaker10() - Baker's order 10 method.

constructHairer10() Hairer's order 10 method.

constructFeagin12Tableau() - Feagin's order 12 method.

constructOno12() - Ono's order 12 method.

constructFeagin14Tableau() Feagin's order 14 method.

Implicit Runge-Kutta Methods

constructImplicitEuler - The 1st order Implicit Euler method.

constructMidpointRule - The 2nd order Midpoint method.

constructTrapezoidalRule - The 2nd order Trapezoidal rule (2nd order LobattoIIIA)

constructLobattoIIIA4 - The 4th order LobattoIIIA

constructLobattoIIIB2 - The 2nd order LobattoIIIB

constructLobattoIIIB4 - The 4th order LobattoIIIB

constructLobattoIIIC2 - The 2nd order LobattoIIIC

constructLobattoIIIC4 - The 4th order LobattoIIIC

constructLobattoIIICStar2 - The 2nd order LobattoIIIC*

constructLobattoIIICStar4 - The 4th order LobattoIIIC*

constructLobattoIIID2 - The 2nd order LobattoIIID

constructLobattoIIID4 - The 4th order LobattoIIID

constructRadauIA3 - The 3rd order RadauIA

constructRadauIA5 - The 5th order RadauIA

constructRadauIIA3 - The 3rd order RadauIIA

constructRadauIIA5 - The 5th order RadauIIA

Tableau Methods

DiffEqDevTools.stability_region — Function

stability_region(z,tab::ODERKTableau)

Calculates the stability function from the tableau at z. Stable if <1.

\[r(z) = 1 + z bᵀ(I - zA)⁻¹ e\]

where e denotes a vector of ones.

stability_region(tab::ODERKTableau; initial_guess=-3.0)

Calculates the length of the stability region in the real axis.

OrdinaryDiffEq.ODE_DEFAULT_TABLEAU — Constant

ODEDEFAULTTABLEAU

Sets the default tableau for the ODE solver. Currently Dormand-Prince 4/5.

Explicit Tableaus

DiffEqDevTools.constructEuler — Function

Euler's method.

DiffEqDevTools.constructRalston — Function

Ralston's Order 2 method.

DiffEqDevTools.constructHeun — Function

Heun's Order 2 method.

DiffEqDevTools.constructKutta3 — Function

Kutta's Order 3 method.

Missing docstring.

Missing docstring for OrdinaryDiffEq.constructBS3. Check Documenter's build log for details.

DiffEqDevTools.constructBogakiShampine3 — Function

constructBogakiShampine3()

Constructs the tableau object for the Bogakai-Shampine Order 2/3 method.

DiffEqDevTools.constructRK4 — Function

Classic RK4 method.

DiffEqDevTools.constructRK438Rule — Function

Classic RK4 3/8's rule method.

DiffEqDevTools.constructRKF4 — Function

Runge-Kutta-Fehberg Order 4/3

DiffEqDevTools.constructRKF5 — Function

Runge-Kutta-Fehlberg Order 4/5 method.

DiffEqDevTools.constructCashKarp — Function

constructCashKarp()

Constructs the tableau object for the Cash-Karp Order 4/5 method.

Missing docstring.

Missing docstring for DiffEqDevTools.constructDormandPrince. Check Documenter's build log for details.

Missing docstring.

Missing docstring for OrdinaryDiffEq.constructBS5. Check Documenter's build log for details.

DiffEqDevTools.constructPapakostasPapaGeorgiou5 — Function

S.N. Papakostas and G. PapaGeorgiou higher error more stable

A Family of Fifth-order Runge-Kutta Pairs, by S.N. Papakostas and G. PapaGeorgiou, Mathematics of Computation,Volume 65, Number 215, July 1996, Pages 1165-1181.

DiffEqDevTools.constructPapakostasPapaGeorgiou52 — Function

S.N. Papakostas and G. PapaGeorgiou less stable lower error Strictly better than DP5

A Family of Fifth-order Runge-Kutta Pairs, by S.N. Papakostas and G. PapaGeorgiou, Mathematics of Computation,Volume 65, Number 215, July 1996, Pages 1165-1181.

DiffEqDevTools.constructTsitouras5 — Function

Runge–Kutta pairs of orders 5(4) using the minimal set of simplifying assumptions, by Ch. Tsitouras, TEI of Chalkis, Dept. of Applied Sciences, GR34400, Psahna, Greece.

DiffEqDevTools.constructLutherKonen5 — Function

Luther and Konen's First Order 5 Some Fifth-Order Classical Runge Kutta Formulas, H.A.Luther and H.P.Konen, Siam Review, Vol. 3, No. 7, (Oct., 1965) pages 551-558.

DiffEqDevTools.constructLutherKonen52 — Function

Luther and Konen's Second Order 5 Some Fifth-Order Classical Runge Kutta Formulas, H.A.Luther and H.P.Konen, Siam Review, Vol. 3, No. 7, (Oct., 1965) pages 551-558.

DiffEqDevTools.constructLutherKonen53 — Function

Luther and Konen's Third Order 5 Some Fifth-Order Classical Runge Kutta Formulas, H.A.Luther and H.P.Konen, Siam Review, Vol. 3, No. 7, (Oct., 1965) pages 551-558.

DiffEqDevTools.constructRungeFirst5 — Function

Runge's First Order 5 method

DiffEqDevTools.constructLawson5 — Function

Lawson's 5th order scheme

An Order Five Runge Kutta Process with Extended Region of Stability, J. Douglas Lawson, Siam Journal on Numerical Analysis, Vol. 3, No. 4, (Dec., 1966) pages 593-597

DiffEqDevTools.constructSharpSmart5 — Function

Explicit Runge-Kutta Pairs with One More Derivative Evaluation than the Minimum, by P.W.Sharp and E.Smart, Siam Journal of Scientific Computing, Vol. 14, No. 2, pages. 338-348, March 1993.

DiffEqDevTools.constructBogakiShampine5 — Function

An Efficient Runge-Kutta (4,5) Pair by P.Bogacki and L.F.Shampine Computers and Mathematics with Applications, Vol. 32, No. 6, 1996, pages 15 to 28

DiffEqDevTools.constructCassity5 — Function

Cassity's Order 5 method

DiffEqDevTools.constructButcher6 — Function

Butcher's First Order 6 method

On Runge-Kutta Processes of High Order, by J. C. Butcher, Journal of the Australian Mathematical Society, Vol. 4, (1964), pages 179 to 194

DiffEqDevTools.constructButcher62 — Function

Butcher's Second Order 6 method

On Runge-Kutta Processes of High Order, by J. C. Butcher, Journal of the Australian Mathematical Society, Vol. 4, (1964), pages 179 to 194

DiffEqDevTools.constructButcher63 — Function

Butcher's Third Order 6

On Runge-Kutta Processes of High Order, by J. C. Butcher, Journal of the Australian Mathematical Society, Vol. 4, (1964), pages 179 to 194

DiffEqDevTools.constructVernerRobust6 — Function

From Verner's Website

DiffEqDevTools.constructTanakaKasugaYamashitaYazaki6A — Function

TanakaKasugaYamashitaYazaki Order 6 A

On the Optimization of Some Eight-stage Sixth-order Explicit Runge-Kutta Method, by M. Tanaka, K. Kasuga, S. Yamashita and H. Yazaki, Journal of the Information Processing Society of Japan, Vol. 34, No. 1 (1993), pages 62 to 74.

DiffEqDevTools.constructTanakaKasugaYamashitaYazaki6B — Function

constructTanakaKasugaYamashitaYazaki Order 6 B

On the Optimization of Some Eight-stage Sixth-order Explicit Runge-Kutta Method, by M. Tanaka, K. Kasuga, S. Yamashita and H. Yazaki, Journal of the Information Processing Society of Japan, Vol. 34, No. 1 (1993), pages 62 to 74.

DiffEqDevTools.constructTanakaKasugaYamashitaYazaki6C — Function

constructTanakaKasugaYamashitaYazaki Order 6 C

On the Optimization of Some Eight-stage Sixth-order Explicit Runge-Kutta Method, by M. Tanaka, K. Kasuga, S. Yamashita and H. Yazaki, Journal of the Information Processing Society of Japan, Vol. 34, No. 1 (1993), pages 62 to 74.

DiffEqDevTools.constructTanakaKasugaYamashitaYazaki6D — Function

constructTanakaKasugaYamashitaYazaki Order 6 D

On the Optimization of Some Eight-stage Sixth-order Explicit Runge-Kutta Method, by M. Tanaka, K. Kasuga, S. Yamashita and H. Yazaki, Journal of the Information Processing Society of Japan, Vol. 34, No. 1 (1993), pages 62 to 74.

DiffEqDevTools.constructHuta6 — Function

Anton Hutas First Order 6 method

Une amélioration de la méthode de Runge-Kutta-Nyström pour la résolution numérique des équations différentielles du premièr ordre, by Anton Huta, Acta Fac. Nat. Univ. Comenian Math., Vol. 1, pages 201-224 (1956).

DiffEqDevTools.constructHuta62 — Function

Anton Hutas Second Order 6 method

Une amélioration de la méthode de Runge-Kutta-Nyström pour la résolution numérique des équations différentielles du premièr ordre, by Anton Huta, Acta Fac. Nat. Univ. Comenian Math., Vol. 1, pages 201-224 (1956).

DiffEqDevTools.constructVerner6 — Function

Verner Order 5/6 method

A Contrast of a New RK56 pair with DP56, by Jim Verner, Department of Mathematics. Simon Fraser University, Burnaby, Canada, 2006.

DiffEqDevTools.constructDormandPrince6 — Function

Dormand-Prince Order 5//6 method

P.J. Prince and J. R. Dormand, High order embedded Runge-Kutta formulae, Journal of Computational and Applied Mathematics . 7 (1981), pp. 67-75.

DiffEqDevTools.constructSharpVerner6 — Function

Sharp-Verner Order 5/6 method

Completely Imbedded Runge-Kutta Pairs, by P. W. Sharp and J. H. Verner, SIAM Journal on Numerical Analysis, Vol. 31, No. 4. (Aug., 1994), pages. 1169 to 1190.

Missing docstring.

Missing docstring for DiffEqDevTools.constructVern6. Check Documenter's build log for details.

DiffEqDevTools.constructClassicVerner6 — Function

EXPLICIT RUNGE-KUTFA METHODS WITH ESTIMATES OF THE LOCAL TRUNCATION ERROR

DiffEqDevTools.constructChummund6 — Function

Chummund's First Order 6 method

A three-dimensional family of seven-step Runge-Kutta methods of order 6, by G. M. Chammud (Hammud), Numerical Methods and programming, 2001, Vol.2, 2001, pages 159-166 (Advanced Computing Scientific journal published by the Research Computing Center of the Lomonosov Moscow State Univeristy)

DiffEqDevTools.constructChummund62 — Function

Chummund's Second Order 6 method

A three-dimensional family of seven-step Runge-Kutta methods of order 6, by G. M. Chammud (Hammud), Numerical Methods and programming, 2001, Vol.2, 2001, pages 159-166 (Advanced Computing Scientific journal published by the Research Computing Center of the Lomonosov Moscow State Univeristy)

DiffEqDevTools.constructPapakostas6 — Function

Papakostas's Order 6

On Phase-Fitted modified Runge-Kutta Pairs of order 6(5), by Ch. Tsitouras and I. Th. Famelis, International Conference of Numerical Analysis and Applied Mathematics, Crete, (2006)

DiffEqDevTools.constructLawson6 — Function

Lawson's Order 6

An Order 6 Runge-Kutta Process with an Extended Region of Stability, by J. D. Lawson, Siam Journal on Numerical Analysis, Vol. 4, No. 4 (Dec. 1967) pages 620-625.

DiffEqDevTools.constructTsitourasPapakostas6 — Function

Tsitouras-Papakostas's Order 6

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

DiffEqDevTools.constructDormandLockyerMcCorriganPrince6 — Function

DormandLockyerMcCorriganPrince Order 6 Global Error Estimation

Global Error estimation with Runge-Kutta triples, by J.R.Dormand, M.A.Lockyer, N.E.McCorrigan and P.J.Prince, Computers and Mathematics with Applications, 18 (1989) pages 835-846.

DiffEqDevTools.constructVernerEfficient6 — Function

From Verner's Website

DiffEqDevTools.constructMikkawyEisa — Function

Mikkawy-Eisa Order 6

A general four-parameter non-FSAL embedded Runge–Kutta algorithm of orders 6 and 4 in seven stages, by M.E.A. El-Mikkawy and M.M.M. Eisa, Applied Mathematics and Computation, Vol. 143, No. 2, (2003) pages 259 to 267.

DiffEqDevTools.constructVernerEfficient7 — Function

From Verner's website

DiffEqDevTools.constructClassicVerner7 — Function

EXPLICIT RUNGE-KUTFA METHODS WITH ESTIMATES OF THE LOCAL TRUNCATION ERROR

DiffEqDevTools.constructSharpVerner7 — Function

Completely Imbedded Runge-Kutta Pairs, by P.W.Sharp and J.H.Verner, Siam Journal on Numerical Analysis, Vol.31, No.4. (August 1994) pages 1169-1190.

DiffEqDevTools.constructTanakaYamashitaStable7 — Function

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

DiffEqDevTools.constructSharpSmart7 — Function

Explicit Runge-Kutta Pairs with One More Derivative Evaluation than the Minimum, by P.W.Sharp and E.Smart, Siam Journal of Scientific Computing, Vol. 14, No. 2, pages. 338-348, March 1993.

DiffEqDevTools.constructTanakaYamashitaEfficient7 — Function

On the Optimization of Some Nine-Stage Seventh-order Runge-Kutta Method, by M. Tanaka, S. Muramatsu and S. Yamashita, Information Processing Society of Japan, Vol. 33, No. 12 (1992) pages 1512-1526.

DiffEqDevTools.constructVernerRobust7 — Function

From Verner's website

Missing docstring.

Missing docstring for OrdinaryDiffEq.constructTanYam7. Check Documenter's build log for details.

DiffEqDevTools.constructEnrightVerner7 — Function

The Relative Efficiency of Alternative Defect Control Schemes for High-Order Continuous Runge-Kutta Formulas W. H. Enright SIAM Journal on Numerical Analysis, Vol. 30, No. 5. (Oct., 1993), pp. 1419-1445.

DiffEqDevTools.constructDormandPrince8 — Function

constructDormandPrice8()

Constructs the tableau object for the Dormand-Prince Order 6/8 method.

DiffEqDevTools.constructRKF8 — Function

constructRKF8()

Constructs the tableau object for the Runge-Kutta-Fehlberg Order 7/8 method.

DiffEqDevTools.constructCooperVerner8 — Function

Some Explicit Runge-Kutta Methods of High Order, by G. J. Cooper and J. H. Verner, SIAM Journal on Numerical Analysis, Vol. 9, No. 3, (September 1972), pages 389 to 405

DiffEqDevTools.constructCooperVerner82 — Function

Some Explicit Runge-Kutta Methods of High Order, by G. J. Cooper and J. H. Verner, SIAM Journal on Numerical Analysis, Vol. 9, No. 3, (September 1972), pages 389 to 405

DiffEqDevTools.constructTsitourasPapakostas8 — Function

Cheap Error Estimation for Runge-Kutta methods, by Ch. Tsitouras and S.N. Papakostas, Siam Journal on Scientific Computing, Vol. 20, Issue 6, Nov 1999.

DiffEqDevTools.constructEnrightVerner8 — Function

The Relative Efficiency of Alternative Defect Control Schemes for High-Order Continuous Runge-Kutta Formulas W. H. Enright SIAM Journal on Numerical Analysis, Vol. 30, No. 5. (Oct., 1993), pp. 1419-1445.

DiffEqDevTools.constructdverk78 — Function

Jim Verner's "Maple" (dverk78)

DiffEqDevTools.constructClassicVerner8 — Function

EXPLICIT RUNGE-KUTFA METHODS WITH ESTIMATES OF THE LOCAL TRUNCATION ERROR

DiffEqDevTools.constructDormandPrince8_64bit — Function

constructDormandPrice8_64bit()

Constructs the tableau object for the Dormand-Prince Order 6/8 method with the approximated coefficients from the paper. This works until below 64-bit precision.

DiffEqDevTools.constructCurtis8 — Function

An Eighth Order Runge-Kutta process with Eleven Function Evaluations per Step, by A. R. Curtis, Numerische Mathematik, Vol. 16, No. 3 (1970), pages 268 to 277

Missing docstring.

Missing docstring for OrdinaryDiffEq.constructTsitPap8. Check Documenter's build log for details.

DiffEqDevTools.constructSharp9 — Function

Journal of Applied Mathematics & Decision Sciences, 4(2), 183-192 (2000), "High order explicit Runge-Kutta pairs for ephemerides of the Solar System and the Moon".

DiffEqDevTools.constructTsitouras9 — Function

Optimized explicit Runge-Kutta pairs of order 9(8), by Ch. Tsitouras, Applied Numerical Mathematics, 38 (2001) 123-134.

DiffEqDevTools.constructTsitouras92 — Function

Optimized explicit Runge-Kutta pairs of order 9(8), by Ch. Tsitouras, Applied Numerical Mathematics, 38 (2001) 123-134.

DiffEqDevTools.constructVernerEfficient9 — Function

From Verner's Webiste

Missing docstring.

Missing docstring for OrdinaryDiffEq.constructVern9. Check Documenter's build log for details.

DiffEqDevTools.constructVerner916 — Function

Verner 1991 First Order 5/6 method

Some Ruge-Kutta Formula Pairs, by J.H.Verner, SIAM Journal on Numerical Analysis, Vol. 28, No. 2 (April 1991), pages 496 to 511.

DiffEqDevTools.constructVerner9162 — Function

Verner 1991 Second Order 5/6 method

Some Ruge-Kutta Formula Pairs, by J.H.Verner, SIAM Journal on Numerical Analysis, Vol. 28, No. 2 (April 1991), pages 496 to 511.

DiffEqDevTools.constructVernerRobust9 — Function

From Verner's Webiste

DiffEqDevTools.constructFeagin10 — Function

Feagin10 in Tableau form

Missing docstring.

Missing docstring for DiffEqDevTools.constructFeagin10Tableau. Check Documenter's build log for details.

DiffEqDevTools.constructOno10 — Function

Ono10

DiffEqDevTools.constructCurtis10 — Function

High-order Explicit Runge-Kutta Formulae, Their uses, and Limitations, A.R.Curtis, J. Inst. Maths Applics (1975) 16, 35-55.

DiffEqDevTools.constructHairer10 — Function

A Runge-Kutta Method of Order 10, E. Hairer, J. Inst. Maths Applics (1978) 21, 47-59.

DiffEqDevTools.constructBaker10 — Function

Tom Baker, University of Teeside. Part of RK-Aid http://www.scm.tees.ac.uk/users/u0000251/research/researcht.htm http://www.scm.tees.ac.uk/users/u0000251/j.r.dormand/t.baker/rk10921m/rk10921m

DiffEqDevTools.constructFeagin12 — Function

Tableau form of Feagin12

DiffEqDevTools.constructOno12 — Function

On the 25 stage 12th order explicit Runge-Kutta method, by Hiroshi Ono. Transactions of the Japan Society for Industrial and applied Mathematics, Vol. 6, No. 3, (2006) pages 177 to 186

Missing docstring.

Missing docstring for DiffEqDevTools.constructFeagin12Tableau. Check Documenter's build log for details.

DiffEqDevTools.constructFeagin14 — Function

Tableau form of Feagin14

Missing docstring.

Missing docstring for DiffEqDevTools.constructFeagin14Tableau. Check Documenter's build log for details.

Implicit Tableaus

DiffEqDevTools.constructImplicitEuler — Function

Implicit Euler Method

DiffEqDevTools.constructMidpointRule — Function

Order 2 Midpoint Method

DiffEqDevTools.constructTrapezoidalRule — Function

Order 2 Trapezoidal Rule (LobattoIIIA2)

DiffEqDevTools.constructLobattoIIIA4 — Function

LobattoIIIA Order 4 method

DiffEqDevTools.constructLobattoIIIB2 — Function

LobattoIIIB Order 2 method

DiffEqDevTools.constructLobattoIIIB4 — Function

LobattoIIIB Order 4 method

DiffEqDevTools.constructLobattoIIIC2 — Function

LobattoIIIC Order 2 method

DiffEqDevTools.constructLobattoIIIC4 — Function

LobattoIIIC Order 4 method

DiffEqDevTools.constructLobattoIIICStar2 — Function

LobattoIIIC* Order 2 method

DiffEqDevTools.constructLobattoIIICStar4 — Function

LobattoIIIC* Order 4 method

DiffEqDevTools.constructLobattoIIID2 — Function

LobattoIIID Order 2 method

DiffEqDevTools.constructLobattoIIID4 — Function

LobattoIIID Order 4 method

DiffEqDevTools.constructGL2 — Function

Gauss-Legendre Order 2.

DiffEqDevTools.constructGL4 — Function

Gauss-Legendre Order 4.

DiffEqDevTools.constructGL6 — Function

Gauss-Legendre Order 6.

DiffEqDevTools.constructRadauIA3 — Function

RadauIA Order 3 method

DiffEqDevTools.constructRadauIA5 — Function

RadauIA Order 5 method

DiffEqDevTools.constructRadauIIA3 — Function

RadauIIA Order 3 method

DiffEqDevTools.constructRadauIIA5 — Function

RadauIIA Order 5 method