ModelingToolkitStandardLibrary: Magnetic Components

Flux Tubes

Flux Tube Utilities

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.PositiveMagneticPort — Function

Positive magnetic port

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.NegativeMagneticPort — Function

Negative magnetic port

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.TwoPort — Function

TwoPort(;name, V_m_start=0.0, Phi_start=0.0)

Partial component with magnetic potential difference between two magnetic ports p and n and magnetic flux Phi from p to n.

Parameters:

V_m_start: Initial magnetic potential difference between both ports

Phi_start: Initial magnetic flux from portp to portn

Basic Flux Tube Blocks

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Ground — Function

Ground(;name)

Zero magnetic potential.

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Idle — Function

Idle(;name)

Idle running branch.

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Short — Function

Short(;name)

Short cut branch.

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.Crossing — Function

Crossing(;name)

Crossing of two branches.

This is a simple crossing of two branches. The ports portp1 and portp2 are connected, as well as portn1 and portn2.

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantPermeance — Function

ConstantPermeance(;name, G_m=1.0)

Constant permeance.

Parameters:

G_m: [H] Magnetic permeance

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantReluctance — Function

ConstantReluctance(;name, R_m=1.0)

Constant reluctance.

Parameters:

R_m: [H^-1] Magnetic reluctance

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.EddyCurrent — Function

EddyCurrent(;name, rho=0.098e-6, l=1, A=1, Phi_start=0.0)

For modelling of eddy current in a conductive magnetic flux tube.

Parameters:

rho: [ohm * m] Resistivity of flux tube material (default: Iron at 20degC)

l: [m] Average length of eddy current path

A: [m^2] Cross sectional area of eddy current path

Phi_start: [Wb] Initial magnetic flux flowing into the port_p

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ElectroMagneticConverter — Function

ElectroMagneticConverter(;name, N, Phi_start=0.0)

Ideal electromagnetic energy conversion.

The electromagnetic energy conversion is given by Ampere's law and Faraday's law respectively V_m = N * i N * dΦ/dt = -v

Parameters:

N: Number of turns

Phi_start: [Wb] Initial magnetic flux flowing into the port_p

Flux Tube Sources

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantMagneticPotentialDifference — Function

Constant magnetomotive force.

Parameters:

V_m: [A] Magnetic potential difference

ModelingToolkitStandardLibrary.Magnetic.FluxTubes.ConstantMagneticFlux — Function

Source of constant magnetic flux.

Parameters:

Phi: [Wb] Magnetic flux