Sampler APIs

Sample

Missing docstring.

Missing docstring for sample. Check Documenter's build log for details.

Samplers

QuasiMonteCarlo.GridSample — Type

GridSample{T}

The grid is given by lb:dx[i]:ub in the ith direction.

QuasiMonteCarlo.UniformSample — Type

struct UniformSample <: SamplingAlgorithm end

Uniformly distributed random numbers.

QuasiMonteCarlo.SobolSample — Type

struct SobolSample <: SamplingAlgorithm end

Samples using the Sobol sequence.

QuasiMonteCarlo.LatinHypercubeSample — Type

struct LatinHypercubeSample <: SamplingAlgorithm end

Samples using a Latin Hypercube.

QuasiMonteCarlo.LatticeRuleSample — Type

struct LatticeRuleSample <: SamplingAlgorithm end

Samples using a randomly-shifted rank-1 lattice rule.

QuasiMonteCarlo.LowDiscrepancySample — Type

struct LowDiscrepancySample{T} <: SamplingAlgorithm

base[i] is the base in the ith direction.

QuasiMonteCarlo.GoldenSample — Type

struct GoldenSample <: SamplingAlgorithm end

QuasiMonteCarlo.KroneckerSample — Type

struct KroneckerSample{A,B} <: SamplingAlgorithm

KroneckerSample(alpha, s0) for a Kronecker sequence, where alpha is an length-d vector of irrational numbers (often sqrt(d)) and s0 is a length-d seed vector (often 0).

QuasiMonteCarlo.SectionSample — Type

struct SectionSample{T} <: SamplingAlgorithm

SectionSample(x0, sampler) where sampler is any sampler above and x0 is a vector of either NaN for a free dimension or some scalar for a constrained dimension.