SciMLProblems

The cornerstone of the SciML common interface is the problem type definition. These definitions are the encoding of mathematical problems into a numerically computable form.

Note About Symbolics and ModelingToolkit

The symbolic analog to the problem interface is the ModelingToolkit AbstractSystem. For example, ODESystem is the symbolic analog to ODEProblem. Each of these system types have a method for constructing the associated problem and function types.

Definition of the SciMLProblem Interface

The following standard principles should be adhered to across all SciMLProblem instantiations.

In-place Specification

Each SciMLProblem type can be called with an "is inplace" (iip) choice. For example:

ODEProblem(f,u0,tspan,p)
ODEProblem{iip}(f,u0,tspan,p)

which is a boolean for whether the function is in the inplace form (mutating to change the first value). This is automatically determined using the methods table but note that for full type-inferrability of the SciMLProblem this iip-ness should be specified.

Additionally, the functions are fully specialized to reduce the runtimes. If one would instead like to not specialize on the functions to reduce compile time, then one can set recompile to false.

Default Parameters

By default, SciMLProblem types use the SciMLBase.NullParameters() singleton to define the absence of parameters by default. The reason is because this throws an informative error if the parameter is used or accessed within the user's function, for example, p[1] will throw an informative error about forgetting to pass parameters.

Keyword Argument Splatting

All SciMLProblem types allow for passing keyword arguments that would get forwarded to the solver. The reason for this is that in many cases, like in EnsembleProblem usage, a SciMLProblem might be associated with some solver configuration, such as a callback or tolerance. Thus, for flexibility the extra keyword arguments to the SciMLProblem are carried to the solver.

problem_type

SciMLProblem types include a non-public API definition of problem_type which holds a trait type corresponding to the way the SciMLProblem was constructed. For example, if a SecondOrderODEProblem constructor is used, the returned problem is simply a ODEProblem for interopability with any ODEProblem algorithm. However, in this case the problem_type will be populated with the SecondOrderODEProblem type, indicating the original definition and extra structure.

Remake

SciMLBase.remake — Function

remake(thing; <keyword arguments>)

Re-construct thing with new field values specified by the keyword arguments.

Problem Traits

SciMLBase.isinplace — Method

isinplace(prob::SciMLProblem)

Determine whether the function of the given problem operates in place or not.

SciMLBase.is_diagonal_noise — Function

is_diagonal_noise(prob::SciMLProblem)

SciMLProblem API

Abstract SciMLProblems

SciMLBase.SciMLProblem — Type

abstract type SciMLProblem

SciMLBase.DEProblem — Type

abstract type DEProblem <: SciMLBase.SciMLProblem

Base type for all DifferentialEquations.jl problems. Concrete subtypes of DEProblem contain the necessary information to fully define a differential equation of the corresponding type.

SciMLBase.AbstractLinearProblem — Type

abstract type AbstractLinearProblem{bType, isinplace} <: SciMLBase.SciMLProblem

Base for types which define linear systems.

SciMLBase.AbstractNonlinearProblem — Type

abstract type AbstractNonlinearProblem{uType, isinplace} <: SciMLBase.DEProblem

Base for types which define nonlinear solve problems (f(u)=0).

Missing docstring.

Missing docstring for SciMLBase.AbstractQuadratureProblem. Check Documenter's build log for details.

SciMLBase.AbstractOptimizationProblem — Type

abstract type AbstractOptimizationProblem{isinplace} <: SciMLBase.SciMLProblem

Base for types which define equations for optimization.

SciMLBase.AbstractNoiseProblem — Type

abstract type AbstractNoiseProblem <: SciMLBase.DEProblem

SciMLBase.AbstractODEProblem — Type

abstract type AbstractODEProblem{uType, tType, isinplace} <: SciMLBase.DEProblem

Base for types which define ODE problems.

SciMLBase.AbstractDiscreteProblem — Type

abstract type AbstractDiscreteProblem{uType, tType, isinplace} <: SciMLBase.AbstractODEProblem{uType, tType, isinplace}

Base for types which define discrete problems.

SciMLBase.AbstractAnalyticalProblem — Type

abstract type AbstractAnalyticalProblem{uType, tType, isinplace} <: SciMLBase.AbstractODEProblem{uType, tType, isinplace}

SciMLBase.AbstractRODEProblem — Type

abstract type AbstractRODEProblem{uType, tType, isinplace, ND} <: SciMLBase.DEProblem

Base for types which define RODE problems.

SciMLBase.AbstractSDEProblem — Type

abstract type AbstractSDEProblem{uType, tType, isinplace, ND} <: SciMLBase.AbstractRODEProblem{uType, tType, isinplace, ND}

Base for types which define SDE problems.

SciMLBase.AbstractDAEProblem — Type

abstract type AbstractDAEProblem{uType, duType, tType, isinplace} <: SciMLBase.DEProblem

Base for types which define DAE problems.

SciMLBase.AbstractDDEProblem — Type

abstract type AbstractDDEProblem{uType, tType, lType, isinplace} <: SciMLBase.DEProblem

Base for types which define DDE problems.

SciMLBase.AbstractConstantLagDDEProblem — Type

abstract type AbstractConstantLagDDEProblem{uType, tType, lType, isinplace} <: SciMLBase.AbstractDDEProblem{uType, tType, lType, isinplace}

SciMLBase.AbstractSecondOrderODEProblem — Type

abstract type AbstractSecondOrderODEProblem{uType, tType, isinplace} <: SciMLBase.AbstractODEProblem{uType, tType, isinplace}

SciMLBase.AbstractBVProblem — Type

abstract type AbstractBVProblem{uType, tType, isinplace} <: SciMLBase.AbstractODEProblem{uType, tType, isinplace}

Base for types which define BVP problems.

SciMLBase.AbstractJumpProblem — Type

abstract type AbstractJumpProblem{P, J} <: SciMLBase.DEProblem

Base for types which define jump problems.

SciMLBase.AbstractSDDEProblem — Type

abstract type AbstractSDDEProblem{uType, tType, lType, isinplace, ND} <: SciMLBase.DEProblem

Base for types which define SDDE problems.

SciMLBase.AbstractConstantLagSDDEProblem — Type

abstract type AbstractConstantLagSDDEProblem{uType, tType, lType, isinplace, ND} <: SciMLBase.AbstractSDDEProblem{uType, tType, lType, isinplace, ND}

SciMLBase.AbstractPDEProblem — Type

abstract type AbstractPDEProblem <: SciMLBase.DEProblem

Base for types which define PDE problems.

Concrete SciMLProblems

Missing docstring.

Missing docstring for LinearProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for NonlinearProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for QuadratureProblem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for OptimizationProblem. Check Documenter's build log for details.

SciMLBase.BVProblem — Type

Defines an BVP problem. Documentation Page: https://diffeq.sciml.ai/stable/types/bvp_types/

Mathematical Specification of a BVP Problem

To define a BVP Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an ODE:

\[\frac{du}{dt} = f(u,p,t)\]

along with an implicit function bc! which defines the residual equation, where

\[bc(u,p,t) = 0\]

is the manifold on which the solution must live. A common form for this is the two-point BVProblem where the manifold defines the solution at two points:

\[u(t_0) = a u(t_f) = b\]

Problem Type

Constructors

TwoPointBVProblem{isinplace}(f,bc!,u0,tspan,p=NullParameters();kwargs...)
BVProblem{isinplace}(f,bc!,u0,tspan,p=NullParameters();kwargs...)

For any BVP problem type, bc! is the inplace function:

bc!(residual, u, p, t)

where residual computed from the current u. u is an array of solution values where u[i] is at time t[i], while p are the parameters. For a TwoPointBVProblem, t = tspan. For the more general BVProblem, u can be all of the internal time points, and for shooting type methods u=sol the ODE solution. Note that all features of the ODESolution are present in this form. In both cases, the size of the residual matches the size of the initial condition.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

Fields

f: The function for the ODE.

bc: The boundary condition function.

u0: The initial condition. Either the initial condition for the ODE as an initial value problem, or a Vector of values for $u(t_i)$ for collocation methods

tspan: The timespan for the problem.

p: The parameters for the problem. Defaults to NullParameters

kwargs: The keyword arguments passed onto the solves.

SciMLBase.DAEProblem — Type

Defines an implicit ordinary differential equation (ODE) or differential-algebraic equation (DAE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/dae_types/

Mathematical Specification of an DAE Problem

To define a DAE Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an ODE:

\[0 = f(du,u,p,t)\]

f should be specified as f(du,u,p,t) (or in-place as f(resid,du,u,p,t)). Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well.

Problem Type

Constructors

DAEProblem(f::DAEFunction,du0,u0,tspan,p=NullParameters();kwargs...)

DAEProblem{isinplace}(f,du0,u0,tspan,p=NullParameters();kwargs...) : Defines the DAE with the specified functions. isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the DiffEqFunctions page.

Fields

f: The function in the ODE.

du0: The initial condition for the derivative.

u0: The initial condition.

tspan: The timespan for the problem.

differential_vars: A logical array which declares which variables are the differential (non algebraic) vars (i.e. du' is in the equations for this variable). Defaults to nothing. Some solvers may require this be set if an initial condition needs to be determined.

p: The parameters for the problem. Defaults to NullParameters

kwargs: The keyword arguments passed onto the solves.

Example Problems

Examples problems can be found in DiffEqProblemLibrary.jl.

To use a sample problem, such as prob_dae_resrob, you can do something like:

#] add DiffEqProblemLibrary
using DiffEqProblemLibrary.DAEProblemLibrary
# load problems
DAEProblemLibrary.importdaeproblems()
prob = DAEProblemLibrary.prob_dae_resrob
sol = solve(prob,IDA())

SciMLBase.DDEProblem — Type

Defines a delay differential equation (DDE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/dde_types/

Mathematical Specification of a DDE Problem

To define a DDE Problem, you simply need to give the function $f$, the initial condition $u_0$ at time point $t_0$, and the history function $h$ which together define a DDE:

\[\frac{du}{dt} = f(u,h,p,t) \qquad (t \geq t_0)\]

\[u(t_0) = u_0,\]

\[u(t) = h(t) \qquad (t < t_0).\]

$f$ should be specified as f(u, h, p, t) (or in-place as f(du, u, h, p, t)), $u_0$ should be an AbstractArray (or number) whose geometry matches the desired geometry of u, and $h$ should be specified as described below. The history function h is accessed for all delayed values. Note that we are not limited to numbers or vectors for $u_0$; one is allowed to provide $u_0$ as arbitrary matrices / higher dimension tensors as well.

Functional Forms of the History Function

The history function h can be called in the following ways:

h(p, t): out-of-place calculation

h(out, p, t): in-place calculation

h(p, t, deriv::Type{Val{i}}): out-of-place calculation of the ith derivative

h(out, p, t, deriv::Type{Val{i}}): in-place calculation of the ith derivative

h(args...; idxs): calculation of h(args...) for indices idxs

Note that a dispatch for the supplied history function of matching form is required for whichever function forms are used in the user derivative function f.

Declaring Lags

Lags are declared separately from their use. One can use any lag by simply using the interpolant of h at that point. However, one should use caution in order to achieve the best accuracy. When lags are declared, the solvers can more efficiently be more accurate and thus this is recommended.

Neutral and Retarded Delay Differential Equations

Note that the history function specification can be used to specify general retarded arguments, i.e. h(p,α(u,t)). Neutral delay differential equations can be specified by using the deriv value in the history interpolation. For example, h(p,t-τ, Val{1}) returns the first derivative of the history values at time t-τ.

Note that algebraic equations can be specified by using a singular mass matrix.

Problem Type

Constructors

DDEProblem(f[, u0], h, tspan[, p]; <keyword arguments>)
DDEProblem{isinplace}(f[, u0], h, tspan[, p]; <keyword arguments>)

Parameter isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the DiffEqFunctions page.

Arguments

f: The function in the DDE.

u0: The initial condition. Defaults to the value h(p, first(tspan)) of the history function evaluated at the initial time point.

h: The history function for the DDE before t0.

tspan: The timespan for the problem.

p: The parameters with which function f is called. Defaults to NullParameters.

constant_lags: A collection of constant lags used by the history function h. Defaults to ().

dependent_lags A tuple of functions (u, p, t) -> lag for the state-dependent lags used by the history function h. Defaults to ().

neutral: If the DDE is neutral, i.e., if delays appear in derivative terms.

order_discontinuity_t0: The order of the discontinuity at the initial time point. Defaults to 0 if an initial condition u0 is provided. Otherwise it is forced to be greater or equal than 1.

kwargs: The keyword arguments passed onto the solves.

Dynamical Delay Differential Equations

Much like Dynamical ODEs, a Dynamical DDE is a Partitioned DDE of the form:

\[\frac{dv}{dt} = f_1(u,t,h) \\ \frac{du}{dt} = f_2(v,h) \\\]

Constructors

DynamicalDDEProblem(f1, f2[, v0, u0], h, tspan[, p]; <keyword arguments>)
DynamicalDDEProblem{isinplace}(f1, f2[, v0, u0], h, tspan[, p]; <keyword arguments>)

Parameter isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Arguments

f: The function in the DDE.

v0 and u0: The initial condition. Defaults to the values h(p, first(tspan))... of the history function evaluated at the initial time point.

h: The history function for the DDE before t0. Must return an object with the indices 1 and 2, with the values of v and u respectively.

tspan: The timespan for the problem.

p: The parameters with which function f is called. Defaults to NullParameters.

constant_lags: A collection of constant lags used by the history function h. Defaults to ().

dependent_lags A tuple of functions (v, u, p, t) -> lag for the state-dependent lags used by the history function h. Defaults to ().

neutral: If the DDE is neutral, i.e., if delays appear in derivative terms.

order_discontinuity_t0: The order of the discontinuity at the initial time point. Defaults to 0 if an initial condition u0 is provided. Otherwise it is forced to be greater or equal than 1.

kwargs: The keyword arguments passed onto the solves.

The for dynamical and second order DDEs, the history function will return an object with the indicies 1 and 2 defined, where h(p, t_prev)[1] is the value of $f_2(v, u, h, p, t_{\mathrm{prev}})$ and h(p, t_prev)[2] is the value of $f_1(v, u, h, p, t_{\mathrm{prev}})$ (this is for consistency with the ordering of the intitial conditions in the constructor). The supplied history function must also return such a 2-index object, which can be accomplished with a tuple (v,u) or vector [v,u].

2nd Order Delay Differential Equations

To define a 2nd Order DDE Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an DDE:

\[u'' = f(u',u,h,p,t)\]

f should be specified as f(du,u,p,t) (or in-place as f(ddu,du,u,p,t)), and u₀ should be an AbstractArray (or number) whose geometry matches the desired geometry of u. Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well.

From this form, a dynamical ODE:

\[v' = f(v,u,h,p,t) \\ u' = v \\\]

Constructors

SecondOrderDDEProblem(f, [, du0, u0], h, tspan[, p]; <keyword arguments>)
SecondOrderDDEProblem{isinplace}(f, [, du0, u0], h, tspan[, p]; <keyword arguments>)

Parameter isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Arguments

f: The function in the DDE.

du0 and u0: The initial condition. Defaults to the values h(p, first(tspan))... of the history function evaluated at the initial time point.

h: The history function for the DDE before t0. Must return an object with the indices 1 and 2, with the values of v and u respectively.

tspan: The timespan for the problem.

p: The parameters with which function f is called. Defaults to NullParameters.

constant_lags: A collection of constant lags used by the history function h. Defaults to ().

dependent_lags A tuple of functions (v, u, p, t) -> lag for the state-dependent lags used by the history function h. Defaults to ().

neutral: If the DDE is neutral, i.e., if delays appear in derivative terms.

order_discontinuity_t0: The order of the discontinuity at the initial time point. Defaults to 0 if an initial condition u0 is provided. Otherwise it is forced to be greater or equal than 1.

kwargs: The keyword arguments passed onto the solves.

As above, the history function will return an object with indices 1 and 2, with the values of du and u respectively. The supplied history function must also match this return type, e.g. by returning a 2-element tuple or vector.

Example Problems

Example problems can be found in DiffEqProblemLibrary.jl.

To use a sample problem, such as prob_ode_linear, you can do something like:

#] add DiffEqProblemLibrary
using DiffEqProblemLibrary.ODEProblemLibrary
# load problems
ODEProblemLibrary.importodeproblems()
prob = ODEProblemLibrary.prob_ode_linear
sol = solve(prob)

SciMLBase.DynamicalDDEProblem — Type

struct DynamicalDDEProblem{iip} <: SciMLBase.AbstractDynamicalDDEProblem

SciMLBase.SecondOrderDDEProblem — Type

struct SecondOrderDDEProblem{iip} <: SciMLBase.AbstractDynamicalDDEProblem

Missing docstring.

Missing docstring for DiscreteProblem. Check Documenter's build log for details.

SciMLBase.NoiseProblem — Type

struct NoiseProblem{N<:SciMLBase.AbstractNoiseProcess, T, K} <: SciMLBase.AbstractNoiseProblem

Missing docstring.

Missing docstring for ODEProblem. Check Documenter's build log for details.

SciMLBase.DynamicalODEProblem — Type

Defines an dynamical ordinary differential equation (ODE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/dynamical_types/

Dynamical ordinary differential equations, such as those arising from the definition of a Hamiltonian system or a second order ODE, have a special structure that can be utilized in the solution of the differential equation. On this page we describe how to define second order differential equations for their efficient numerical solution.

Mathematical Specification of a Dynamical ODE Problem

These algorithms require a Partitioned ODE of the form:

\[\frac{dv}{dt} = f_1(u,t) \\ \frac{du}{dt} = f_2(v) \\\]

This is a Partitioned ODE partitioned into two groups, so the functions should be specified as f1(dv,v,u,p,t) and f2(du,v,u,p,t) (in the inplace form), where f1 is independent of v (unless specified by the solver), and f2 is independent of u and t. This includes discretizations arising from SecondOrderODEProblems where the velocity is not used in the acceleration function, and Hamiltonians where the potential is (or can be) time-dependent but the kinetic energy is only dependent on v.

Note that some methods assume that the integral of f2 is a quadratic form. That means that f2=v'*M*v, i.e. $\int f_2 = \frac{1}{2} m v^2$, giving du = v. This is equivalent to saying that the kinetic energy is related to $v^2$. The methods which require this assumption will lose accuracy if this assumption is violated. Methods listed make note of this requirement with "Requires quadratic kinetic energy".

Constructor

DynamicalODEProblem(f::DynamicalODEFunction,v0,u0,tspan,p=NullParameters();kwargs...)
DynamicalODEProblem{isinplace}(f1,f2,v0,u0,tspan,p=NullParameters();kwargs...)

Defines the ODE with the specified functions. isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

Fields

f1 and f2: The functions in the ODE.

v0 and u0: The initial conditions.

tspan: The timespan for the problem.

p: The parameters for the problem. Defaults to NullParameters

kwargs: The keyword arguments passed onto the solves.

SciMLBase.SecondOrderODEProblem — Type

Defines a second order ordinary differential equation (ODE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/dynamical_types/

Mathematical Specification of a 2nd Order ODE Problem

To define a 2nd Order ODE Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an ODE:

\[u'' = f(u',u,p,t)\]

f should be specified as f(du,u,p,t) (or in-place as f(ddu,du,u,p,t)), and u₀ should be an AbstractArray (or number) whose geometry matches the desired geometry of u. Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well.

From this form, a dynamical ODE:

\[v' = f(v,u,p,t) \\ u' = v \\\]

is generated.

Constructors

SecondOrderODEProblem{isinplace}(f,du0,u0,tspan,callback=CallbackSet())

Defines the ODE with the specified functions.

Fields

f: The function for the second derivative.

du0: The initial derivative.

u0: The initial condition.

tspan: The timespan for the problem.

callback: A callback to be applied to every solver which uses the problem. Defaults to nothing.

SciMLBase.SplitODEProblem — Type

Defines a split ordinary differential equation (ODE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/splitodetypes/

Mathematical Specification of a Split ODE Problem

To define a SplitODEProblem, you simply need to give a two functions $f_1$ and $f_2$ along with an initial condition $u_0$ which define an ODE:

\[\frac{du}{dt} = f_1(u,p,t) + f_2(u,p,t)\]

f should be specified as f(u,p,t) (or in-place as f(du,u,p,t)), and u₀ should be an AbstractArray (or number) whose geometry matches the desired geometry of u. Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well.

Many splits are at least partially linear. That is the equation:

\[\frac{du}{dt} = Au + f_2(u,p,t)\]

For how to define a linear function A, see the documentation for the DiffEqOperators.

Constructors

SplitODEProblem(f::SplitFunction,u0,tspan,p=NullParameters();kwargs...)
SplitODEProblem{isinplace}(f1,f2,u0,tspan,p=NullParameters();kwargs...)

The isinplace parameter can be omitted and will be determined using the signature of f2. Note that both f1 and f2 should support the in-place style if isinplace is true or they should both support the out-of-place style if isinplace is false. You cannot mix up the two styles.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

Under the hood, a SplitODEProblem is just a regular ODEProblem whose f is a SplitFunction. Therefore you can solve a SplitODEProblem using the same solvers for ODEProblem. For solvers dedicated to split problems, see Split ODE Solvers.

For specifying Jacobians and mass matrices, see the DiffEqFunctions page.

Fields

f1, f2: The functions in the ODE.

u0: The initial condition.

tspan: The timespan for the problem.

p: The parameters for the problem. Defaults to NullParameters

kwargs: The keyword arguments passed onto the solves.

SciMLBase.IncrementingODEProblem — Type

Experimental

SciMLBase.RODEProblem — Type

Defines a random ordinary differential equation (RODE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/rode_types/

Mathematical Specification of a RODE Problem

To define a RODE Problem, you simply need to give the function $f$ and the initial condition $u_0$ which define an ODE:

\[\frac{du}{dt} = f(u,p,t,W(t))\]

where W(t) is a random process. f should be specified as f(u,p,t,W) (or in-place as f(du,u,p,t,W)), and u₀ should be an AbstractArray (or number) whose geometry matches the desired geometry of u. Note that we are not limited to numbers or vectors for u₀; one is allowed to provide u₀ as arbitrary matrices / higher dimension tensors as well.

Constructors

RODEProblem(f::RODEFunction,u0,tspan,p=NullParameters();noise=WHITE_NOISE,rand_prototype=nothing,callback=nothing)

RODEProblem{isinplace}(f,u0,tspan,p=NullParameters();noise=WHITE_NOISE,rand_prototype=nothing,callback=nothing,mass_matrix=I) : Defines the RODE with the specified functions. The default noise is WHITE_NOISE. isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the DiffEqFunctions page.

Fields

f: The drift function in the SDE.

u0: The initial condition.

tspan: The timespan for the problem.

p: The optional parameters for the problem. Defaults to NullParameters.

noise: The noise process applied to the noise upon generation. Defaults to Gaussian white noise. For information on defining different noise processes, see the noise process documentation page

rand_prototype: A prototype type instance for the noise vector. It defaults to nothing, which means the problem should be interpreted as having a noise vector whose size matches u0.

kwargs: The keyword arguments passed onto the solves.

SciMLBase.SDDEProblem — Type

Defines a stochastic delay differential equation (SDDE) problem. Documentation Page: https://diffeq.sciml.ai/stable/types/sdde_types/

Mathematical Specification of a Stochastic Delay Differential Equation (SDDE) Problem

To define a SDDE Problem, you simply need to give the drift function $f$, the diffusion function g, the initial condition $u_0$ at time point $t_0$, and the history function $h$ which together define a SDDE:

\[du = f(u,h,p,t)dt + g(u,h,p,t)dW_t \qquad (t \geq t_0)\]

\[u(t_0) = u_0,\]

\[u(t) = h(t) \qquad (t < t_0).\]

$f$ should be specified as f(u, h, p, t) (or in-place as f(du, u, h, p, t)) (and $g$ should match). $u_0$ should be an AbstractArray (or number) whose geometry matches the desired geometry of u, and $h$ should be specified as described below. The history function h is accessed for all delayed values. Note that we are not limited to numbers or vectors for $u_0$; one is allowed to provide $u_0$ as arbitrary matrices / higher dimension tensors as well.

Note that this functionality should be considered experimental.

Functional Forms of the History Function

The history function h can be called in the following ways:

h(p, t): out-of-place calculation

h(out, p, t): in-place calculation

h(p, t, deriv::Type{Val{i}}): out-of-place calculation of the ith derivative

h(out, p, t, deriv::Type{Val{i}}): in-place calculation of the ith derivative

h(args...; idxs): calculation of h(args...) for indices idxs

Note that a dispatch for the supplied history function of matching form is required for whichever function forms are used in the user derivative function f.

Declaring Lags

Lags are declared separately from their use. One can use any lag by simply using the interpolant of h at that point. However, one should use caution in order to achieve the best accuracy. When lags are declared, the solvers can more efficiently be more accurate and thus this is recommended.

Neutral, Retarded, and Algebraic Stochastic Delay Differential Equations

Note that the history function specification can be used to specify general retarded arguments, i.e. h(p,α(u,t)). Neutral delay differential equations can be specified by using the deriv value in the history interpolation. For example, h(p,t-τ, Val{1}) returns the first derivative of the history values at time t-τ.

Note that algebraic equations can be specified by using a singular mass matrix.

Problem Type

Constructors

SDDEProblem(f,g[, u0], h, tspan[, p]; <keyword arguments>)
SDDEProblem{isinplace}(f,g[, u0], h, tspan[, p]; <keyword arguments>)

Parameter isinplace optionally sets whether the function is inplace or not. This is determined automatically, but not inferred.

Parameters are optional, and if not given then a NullParameters() singleton will be used which will throw nice errors if you try to index non-existent parameters. Any extra keyword arguments are passed on to the solvers. For example, if you set a callback in the problem, then that callback will be added in every solve call.

For specifying Jacobians and mass matrices, see the DiffEqFunctions page.

Arguments

f: The drift function in the SDDE.

g: The diffusion function in the SDDE.

u0: The initial condition. Defaults to the value h(p, first(tspan)) of the history function evaluated at the initial time point.

h: The history function for the DDE before t0.

tspan: The timespan for the problem.

p: The parameters with which function f is called. Defaults to NullParameters.

constant_lags: A collection of constant lags used by the history function h. Defaults to ().

dependent_lags A tuple of functions (u, p, t) -> lag for the state-dependent lags used by the history function h. Defaults to ().

neutral: If the DDE is neutral, i.e., if delays appear in derivative terms.

order_discontinuity_t0: The order of the discontinuity at the initial time point. Defaults to 0 if an initial condition u0 is provided. Otherwise it is forced to be greater or equal than 1.

kwargs: The keyword arguments passed onto the solves.

SciMLBase.SplitSDEProblem — Type

struct SplitSDEProblem{iip} <: SciMLBase.AbstractSplitSDEProblem

SciMLBase.DynamicalSDEProblem — Type

struct DynamicalSDEProblem{iip} <: SciMLBase.AbstractDynamicalSDEProblem

Missing docstring.

Missing docstring for SteadyStateProblem. Check Documenter's build log for details.

SciMLBase.PDEProblem — Type

struct PDEProblem{P, E, S} <: SciMLBase.AbstractPDEProblem